refactor: front-end refactor container
