+++ app/components/BlockDescription/BlockDescription.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { ipcRenderer } from 'electron';
-
+import './blockDescription.scss'
 export class BlockDescription extends Component {
     constructor(props) {
         super(props);
