+++ app/components/BlockDescription/blockDescription.scss
@@ -18,19 +18,8 @@
   font-size: 16px;
   word-wrap: break-word;
 }
-.inputField_title{
-  width:30%;
-  height: 60%;
-  border: 0;
-  border-radius: 4px;
-  background-color: rgba(189, 212, 203, 0.61);
-  color: #143938;
-  font-size: 22px;
-  word-wrap: break-word;
-  text-align: center;
-}
-.inputField_title:focus{
-  outline: none;
+.inputField:focus{
+  outline: 0;
 }
 .descriptionInput::placeholder{
   color:#1e525054;
