+++ app/containers/Collection/BlockContainer.jsx
@@ -2,22 +2,22 @@
 import React, { Component } from 'react';
 
 // import ImageBlock from '../components/ImageBlock';
-import ImageBlock from '../components/Block/ImageBlock';
-import VideoBlock from "../components/Block/VideoBlock";
-import TextBlock from "../components/Block/TextBlock";
-import AudioBlock from "../components/Block/AudioBlock";
+import ImageBlock from '../../components/Block/ImageBlock';
+import VideoBlock from "../../components/Block/VideoBlock";
+import TextBlock from "../../components/Block/TextBlock";
+import AudioBlock from "../../components/Block/AudioBlock";
 
 // Electron module
 import { ipcRenderer } from "electron";
 
 // Third party module
-import { AudioRecorder } from '../media-capturer/audio-recorder';
-import { videoRecordStart, videoRecordStop } from '../media-capturer/video-recorder';
-import { JSONManager } from "../renderer/json-manager";
-import { NoteManager } from "../renderer/note-manager";
+import { AudioRecorder } from '../../media-capturer/audio-recorder';
+import { videoRecordStart, videoRecordStop } from '../../media-capturer/video-recorder';
+import { JSONManager } from "../../renderer/json-manager";
+import { NoteManager } from "../../renderer/note-manager";
 
 // Import media API modules
-import { getScreenshot } from '../media-capturer/screenshot';
+import { getScreenshot } from '../../media-capturer/screenshot';
 
 // Notification
 import Noty from 'noty';
