+++ app/containers/Collection/collection.scss
@@ -1,0 +1,162 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: 'Ubuntu', 'Noto Sans TC', 'Noto Sans SC', serif;
+  /*Latin font:ubuntu, Chinese font: Noto, others: sans-serif */
+}
+::-webkit-scrollbar {
+  display: none;
+}
+
+.App {
+  font-size: 0;
+  display:flex;
+  flex-flow: column wrap;
+}
+
+/* blocks */
+.allBlocks{
+  font-size: 16px;
+  margin-bottom: 25px;
+  padding: 10px 0;
+  display:inline-flexbox;
+  flex-flow: column nowrap;
+  /* border: 2px solid rgb(255, 175, 55); */
+}
+
+/*header style start*/
+header {
+  width: 100%;
+
+  background-color: #143938;
+  height: 30px;
+  top: 0;
+  position: sticky;
+  z-index: 298;
+}
+
+.file-title {
+  text-align: center;
+  padding: 2px;
+  color: #fff;
+  font-weight: 300;
+  font-size: 22px;
+}
+
+.inputField_title{
+  width:30%;
+  height: 60%;
+  border: 0;
+  border-radius: 4px;
+  background-color: rgba(189, 212, 203, 0.61);
+  color: #143938;
+  font-size: 22px;
+  word-wrap: break-word;
+  text-align: center;
+}
+.inputField_title:focus{
+  outline: none;
+}
+/*header style end*/
+
+/*block ui override*/
+.block-ui-overlay {
+  position: fixed;
+}
+
+.loading-indicator {
+  display: none;
+}
+
+/* progressBar  */
+.progressContainer {
+  width: 8px;
+  height: 100%;
+  margin: 20px 0 0 calc(110px + 7%);
+  background-color: #bdd4cb;
+  position: fixed;
+}
+
+.progressBar {
+  width: 8px;
+  background: linear-gradient(135deg, #69cb43, #35c189 60%, #35c189);
+}
+
+/* progressBar end */
+
+
+/*navBar*/
+.navigatorContainer{
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  background-color: #143938;
+  height: 30px;
+  bottom: 0;
+  position: fixed;
+  z-index: 300;
+}
+.viewMode{
+  width: 70%;
+  display: flex;
+  justify-content:space-around;
+}
+.selectMode{
+  display: none;
+  width: 70%;
+  justify-content:space-around;
+}
+
+.navigationBar img {
+  width: 16px;
+
+  cursor: pointer;  
+  color:#fafafa;
+}
+
+.search{
+  padding-top: 5px;
+  padding-left: 5%;
+}
+.search > div {
+  display: inline-block;
+  position: relative;
+}
+
+.search > div:after {
+  content: "";
+  background: white;
+  width: 3px;
+  height: 6px;
+  position: absolute;
+  top: 13px;
+  right: -1px;
+  transform: rotate(135deg);
+  box-shadow: 1px 0 #272a2b;
+}
+
+.search > div > input {
+  font-size: 12px;
+  font-weight: 100;
+  background: rgb(255, 255, 255);
+  width: 5px;
+  height: 5px;
+  padding: 6px;
+  border: solid 2px white;
+  outline: none;
+  border-radius: 10px;
+  box-shadow: 0 1px #272a2b;
+  transition: width 0.5s;
+}
+
+.search > div > input::placeholder {
+  color: #1e525054;
+  opacity: 1;
+}
+.search > div > input:focus,
+.search > div > input:valid {
+  width: 180px;
+}
+
+/* navBar end */
