+++ app/containers/Home/home.scss
@@ -1,0 +1,296 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+html,
+body {
+  height: 100%;
+  overflow: hidden;
+  font-family: 'Roboto Condensed', sans-serif;
+}
+
+.page {
+  height: 100%;
+  background-color: #143938;
+}
+
+.content {
+  height: 200vh;
+  transform-origin: top left;
+  transition: transform 0.7s cubic-bezier(1, 0.005, 0.24, 1);
+  background-color: #f0f0f0;
+}
+
+.content_inner {
+  height: 100%;
+  overflow-y: auto;
+  padding: 50px 20%;
+  text-align: center;
+}
+.user_login{
+  display: flex;
+  flex-flow: column nowrap;
+  margin-bottom: 20px;
+}
+.user_login button{
+  outline: none !important;
+  outline-offset: none !important;
+  border: 1px solid transparent;
+  background-color: transparent;
+}
+.user_login img{
+  width: 30px;
+}
+
+.menu_toggle {
+  z-index: 900;
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: block;
+  cursor: pointer;
+  width: 100px;
+  height: 80px;
+  background-color: #143938;
+  border-bottom-right-radius: 100%;
+}
+
+.menu_toggle:active i {
+  opacity: 0.8;
+}
+
+.menu_toggle i {
+  color: #f0f0f0;
+}
+
+.menu_toggle .menu_open,
+.menu_toggle .menu_close {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  margin-top: -15px;
+  margin-left: -12px;
+  transition: transform 0.7s cubic-bezier(1, 0.005, 0.24, 1);
+}
+
+.menu_toggle .menu_open {
+  transform-origin: -100px -100px;
+}
+
+.menu_toggle .menu_close {
+  transform: rotate(20deg);
+  transform-origin: -100px -160px;
+}
+
+.menu_items {
+  position: fixed;
+  bottom: 0;
+  left: 50px;
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+.menu_items li {
+  padding-bottom: 10px;
+  padding-right: 30px;
+  height: 60px;
+  margin-bottom: 30px;
+  transform: translateX(-300px);
+  transition: transform 0.7s 0s cubic-bezier(1, 0.005, 0.24, 1);
+}
+
+.menu_items li:nth-child(2) {
+  margin-left: 40px;
+}
+
+.menu_items li:nth-child(3) {
+  margin-left: 80px;
+}
+
+.menu_items a {
+  display: block;
+  text-decoration: none;
+  text-transform: uppercase;
+  letter-spacing: 2px;
+  color: #BDD4CB;
+  transition: color 0.2s;
+}
+
+.menu_items a .icon {
+  position: relative;
+  display: inline-block;
+  margin-right: 25px;
+  color: #f0f0f0;
+  font-size: 30px;
+}
+
+.menu_items a .icon:after {
+  position: relative;
+  top: 50%;
+  left: 50%;
+  content: '';
+  text-align: center;
+  display: block;
+  width: 60px;
+  height: 60px;
+  margin-left: -32px;
+  margin-top: -45px;
+  border-radius: 50%;
+  border: 2px solid #f0f0f0;
+  transition: border-color 0.2s;
+}
+
+.menu_items a:hover {
+  color: #f0f0f0;
+}
+
+.menu_items a:hover .icon:after {
+  border-color: #69CB43;
+}
+
+.menu_items a:active .icon {
+  color: #69CB43;
+}
+
+.shazam .content {
+  transform: rotate(-30deg);
+}
+
+.shazam .menu_open {
+  transform: rotate(-20deg);
+}
+
+.shazam .menu_close {
+  transform: rotate(0);
+}
+
+.shazam .menu_items li {
+  transform: translateX(0);
+  transition: transform 0.35s 0.45s cubic-bezier(0.175, 0.885, 0.32, 1.275);
+}
+
+.shazam .menu_items li:nth-child(2) {
+  transition-delay: 0.47s;
+}
+
+.shazam .menu_items li:nth-child(3) {
+  transition-delay: 0.48s;
+}
+
+h1 {
+  color: #35C189;
+  text-align: center;
+}
+
+h2{
+  color: #35C189;
+  text-align: left;
+  margin-top: 40px;
+  padding-bottom: 15px;
+  border-bottom: 1px solid #143938;
+  outline: none;
+}
+
+.content_search{
+  background: #143938;
+  border-radius: 5px;
+  padding: 5px;
+  vertical-align: text-bottom;
+  width: 390px;
+  margin: 0px auto;
+}
+
+.search_bar {
+  width: 300px;
+  height: 35px;
+  border: 0px;
+  border-radius: 5px;
+  padding-left: 5px;
+  color: #BDD4CB;
+  outline: none;
+}
+
+.search_bar:focus{
+  color: #143938;
+}
+
+.search_icon{
+  background: white;
+  color: #143938;
+  padding: 5px;
+  border-radius: 4px;
+  margin-left: 5px;
+  font-size: 25px;
+  cursor: pointer;
+  display: inline-block;
+  vertical-align: top;
+}
+
+.content_inner .icon {
+  padding: 10px;
+}
+
+.btn {
+  color:#143938;
+  /* border-radius: 10px;
+  margin: 20px;
+  width: 150px;
+  height: 150px;
+  border: 1px solid #35C189; */
+}
+.btn:hover{
+  color: #35C189;
+  
+}
+.btn:focus{
+  outline: none !important;
+  box-shadow: none;
+}
+
+.open_recent_btn:disabled{
+  color: #14393888;
+}
+
+.open_recent_btn{
+  background-color: #f0f0f0;
+  border: none;
+  font-size: 30px;
+  color: #143938;
+  padding-right: 10px;
+  position: relative;
+  right: -70%;
+  outline: none !important;
+  outline-offset: none !important;
+}
+
+.visible {
+  visibility: visible;
+  opacity: 1;
+  transition-duration: 0.3s;
+}
+
+.hidden {
+  visibility: hidden;
+  opacity: 0;  
+  transition-duration: 0.3s;
+}
+
+.open_rotate{
+  transform: rotate(180deg);
+  transition-duration: 0.3s;
+}
+
+.close_rotate{
+  transform: rotate(0deg);
+  transition-duration: 0.3s;
+}
+
+.pop_trigger{
+  margin: 20px;
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+}
