--- app/containers/css/ControlBar.css
@@ -1,63 +1,0 @@
-body {
-    background-color: rgb(68, 68, 68);
-    color: white;
-    -webkit-user-select: none;
-    -webkit-app-region: drag;
-}
-
-input{
-    border-radius: 5px;
-    padding-left: 10px;
-    border-color: white;
-    border: 1px;
-    background-color: rgb(90, 90, 90);
-    color: white;
-    -webkit-app-region: no-drag;
-    margin-bottom: 10px;
-    outline: none;
-}
-
-textarea {
-    border-radius: 5px;
-    padding-left: 10px;
-    background-color: rgb(90, 90, 90);
-    color: white;
-    -webkit-app-region: no-drag;
-    outline: none;
-}
-
-::-webkit-scrollbar{
-    display: none;
-}
-
-.textdiv {
-    padding: 10px;
-}
-
-.textbtndiv {
-    position: absolute;
-    right: 20px;
-}
-
-.btn {
-    background-color: rgb(68, 68, 68);
-    -webkit-app-region: no-drag;
-    padding-right: 1px;
-    padding-top: 0px;
-    padding-bottom: 0px;
-}
-.btn:focus{
-    outline: none !important;
-    box-shadow: none;
-}
-
-.icon {
-    background-color: rgb(68, 68, 68);
-    width: 20px;
-    height: 20px;
-}
-
-.bar_container {
-    padding-top: 5px;
-    padding-bottom: 5px;
-}
