+++ app/indexCollection.jsx
@@ -2,6 +2,6 @@ import "@babel/polyfill";
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Collection from './containers/Collection'
+import Collection from './containers/Collection/Collection'
 
 ReactDOM.render(<Collection />, document.getElementById('root'));
